requirements.txt
python-dotenv>=1.0.1
langchain>=0.2.17
langchain-google-genai>=2.1.9
faiss-cpu>=1.8.0.post1
tiktoken>=0.7.0
google-genai>=0.7.0

# optional niceties
rich>=13.7.1
requests>=2.31.0
beautifulsoup4>=4.13.5
# uvicorn>=0.23.2
# fastapi>=0.100.0
